// ui/admin/src/models/agentView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wanaku API
 * Wanaku MCP proxy management API
 * OpenAPI spec version: 0.3.0
 */

export interface AgentView {
  address: string;
  /** @nullable */
  cardAddress?: string | null;
  description: string;
  name: string;
  namespace: string;
  proxyUrl: string;
}